Add a new metric that tracks the coordinator changes per cluster
